refactor: extract create capability command
